+++ packages/@expo/cli/src/export/exportApp.ts
@@ -136,6 +136,7 @@ export async function exportAppAsync(
     // split. Hence, there's another separate `copyPublicFolderAsync` call below for `web`
     await copyPublicFolderAsync(publicPath, outputPath);
 
+    let templateHtml: string | undefined;
     // Can be empty during web-only SSG.
     if (spaPlatforms.length) {
       await Promise.all(
@@ -239,6 +240,9 @@ export async function exportAppAsync(
               html = modifyHtml(html);
             }
 
+            // HACK: This is used for adding SSR shims in React Server Components.
+            templateHtml = html;
+
             // Generate SPA-styled HTML file.
             // If web exists, then write the template HTML file.
             files.set('index.html', {
@@ -256,6 +260,7 @@ export async function exportAppAsync(
             files,
             platform: 'web',
             apiRoutesOnly: !isWeb,
+            templateHtml,
           });
         }
       }